HackCybersec/team1/osint_analyzer_v1.2/requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dateutil==2.8.2
python-whois==0.8.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==2.3.7
